refactor: add service package and change some classes diretory
- service package 를 만들고 Judge 와 Racing class 의 이름을 변경한 뒤 package 에 추가하였다.
