+++ src/main/java/racingcar/RacingCarGame.java
@@ -4,6 +4,8 @@ import java.util.ArrayList;
 
 import racingcar.receiver.NameReceiver;
 import racingcar.receiver.TryCountReceiver;
+import racingcar.service.JudgeService;
+import racingcar.service.RacingService;
 import racingcar.util.View;
 
 public class RacingCarGame {
@@ -21,12 +23,12 @@ public class RacingCarGame {
 
 		// 레이싱 실행
 		View.printRacingStartView();
-		Racing racing = new Racing();
-		ArrayList<Car> carList = racing.start(nameList, tryCount);
-		View.printRacingLogView(racing.getRacingLog());
+		RacingService racingService = new RacingService();
+		ArrayList<Car> carList = racingService.start(nameList, tryCount);
+		View.printRacingLogView(racingService.getRacingLog());
 
 		// 최종 우승자 발표
-		ArrayList<String> winnerList = new Judge().makeWinnerList(carList);
+		ArrayList<String> winnerList = new JudgeService().makeWinnerList(carList);
 		View.printWinnerView(winnerList);
 	}
 }
